requirements.txt
openai>=1.5.0
httpcore>=0.14.0
google-genai